🚀 chore(.env): add DATABASE_URL environment variable 🚀 chore(.vscode/settings.json): add WillLuke.nextjs.addTypesOnSave and WillLuke.nextjs.hasPrompted settings ✨ feat(login/page.tsx): create login page component ✨ feat(register/page.tsx): create register page component ✨ feat(api/auth/[...nextauth]/route.ts): create authentication route ✨ feat(globals.css): add styles for ag-grid and webkit-scrollbar ✨ feat(services/firebase.ts): add firebase configuration and initialize firebase app ✨ feat(icons.tsx): add Google icon component ✨ feat(login-form.tsx): create login form component with email, password, and Google login functionality The changes add new components and functionality to the application, including a login page, register page, authentication route, firebase configuration, and a login form component with email, password, and Google login functionality. Additionally, styles
🚀 feat(signup-form.tsx): create SignupForm component
✨ feat(ui/button.tsx): add 'google' variant to Button component
✨ feat(ui/card.tsx): create Card component
✨ feat(ui/input-field.tsx): create Input component
✨ feat(auth/useLogin.tsx): create useLogin hook
✨ feat(auth/useSignup.tsx): create useSignup hook
📦 chore(package.json): add zod, boring-avatars, dotenv, firebase-admin, framer-motion, prisma, next-auth, and next-auth/firebase-adapter
The changes add new components, hooks, and dependencies to the project. The SignupForm component is created to handle user registration. The Button component is updated to include a 'google' variant. The Card and Input components are created to be used in the SignupForm component. The useLogin and useSignup hooks are created to handle user authentication. The zod, boring-

✨ feat(prisma): add Prisma schema file
This commit adds a new file, `schema.prisma`, which defines the data model for the application using Prisma. The schema defines the following models: Account, Session, User, Role, TaskBoard, Task, Comment, Notification, Attachment, and two enums: TaskStatus and TaskPriority. The schema also defines the generator and datasource blocks for Prisma.